refactor: 챌린지 관련 로직을 훅으로 분리 (useChallengeProofCheck, useChallengeComplete, useChallengeConfirm, useChallengeSelect)
