+++ src/app/challenge/page.tsx
@@ -2,32 +2,15 @@
 import ChallengeHeader from '@/components/challenge/ChallengeHeader';
 import ChallengeList from '@/components/challenge/ChallengeList';
 import BottomSheet from '@/components/ui/bottomsheet/BottomSheet';
-import useBottomSheetStore from '@/stores/useBottomSheetStore';
-import { Challenge } from '@/types/challenge.types';
-import { useState } from 'react';
 import ChallengeDetail from '@/components/challenge/ChallengeDetail';
 import useChallenge from '@/hooks/queries/useChallenge';
-import { useQueryClient } from '@tanstack/react-query';
-import { QUERY_KEYS } from '@/hooks/queries/queryKeys';
-import { updateCurrentChallenge } from '@/services/challenge/updateChallenges';
+import { useChallengeSelect } from '@/hooks/queries/useChallengeSelect';
+import { useChallengeConfirm } from '@/hooks/queries/useChallengeConfirm';
 
 const ChallengePage = () => {
-  const { close } = useBottomSheetStore();
-  const [selectedChallenge, setSelectedChallenge] = useState<Challenge | null>(null);
   const { data, isLoading } = useChallenge();
-  const queryClient = useQueryClient();
-
-  const onSelectChallenge = (id: number) => {
-    const selectChallenge = data?.challenges?.find((challenge) => challenge.id === id) ?? null;
-    setSelectedChallenge(selectChallenge);
-  };
-
-  const handleConfirmChallenge = async () => {
-    if (!selectedChallenge) return;
-    await updateCurrentChallenge(selectedChallenge.id);
-    queryClient.invalidateQueries({ queryKey: QUERY_KEYS.challenge() });
-    close();
-  };
+  const { selectedChallenge, onSelectChallenge } = useChallengeSelect(data?.challenges ?? null);
+  const { handleConfirmChallenge } = useChallengeConfirm(selectedChallenge);
 
   if (isLoading || !data) return '로딩 중입니다...';
